requirements.txt
openai==0.27.0
python-telegram-bot==20.0
